Move Shape helper functions into Shape and refactor AudioArgs
